backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.10.0
python-dotenv==1.0.1
alembic==1.14.0
httpx==0.27.2
